use size_t & ssize_t

+++ include/msg_queue.h
@@ -34,7 +34,7 @@ struct swMsgQueue {
 int swMsgQueue_create(swMsgQueue *q, int blocking, key_t msg_key, int perms);
 void swMsgQueue_set_blocking(swMsgQueue *q, uint8_t blocking);
 int swMsgQueue_set_capacity(swMsgQueue *q, int queue_bytes);
-int swMsgQueue_push(swMsgQueue *q, swQueue_data *in, int data_length);
-int swMsgQueue_pop(swMsgQueue *q, swQueue_data *out, int buffer_length);
+ssize_t swMsgQueue_push(swMsgQueue *q, swQueue_data *in, size_t data_length);
+ssize_t swMsgQueue_pop(swMsgQueue *q, swQueue_data *out, size_t buffer_length);
 int swMsgQueue_stat(swMsgQueue *q, int *queue_num, int *queue_bytes);
 int swMsgQueue_free(swMsgQueue *q);
